Allow for decimals in years of service

+++ pensions/models.py
@@ -91,7 +91,7 @@ class Benefit(VintagedModel):
     last_name = models.CharField(max_length=256)
     full_name = models.CharField(max_length=256*2)
     amount = models.DecimalField(max_digits=10, decimal_places=2)
-    years_of_service = models.IntegerField(null=True, blank=True)
+    years_of_service = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     final_salary = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     start_date = models.DateField(null=True, blank=True)
     status = models.CharField(max_length=256, null=True, blank=True)
